move schedule routes to own file, require in app
