+++ app.js
@@ -29,12 +29,12 @@ app.locals.appTitle = `${capitalize(projectName)}`;
 
 const indexRoute = require("./routes/index.routes");
 app.use("/", indexRoute);
-
-
 const authRoutes = require("./routes/auth.routes");
 app.use("/", authRoutes);
 const activityRoutes = require("./routes/activity.routes");
 app.use("/", activityRoutes);
+const scheduleRoutes = require("./routes/schedule.routes");
+app.use("/", scheduleRoutes);
 
 // ❗ To handle errors. Routes that don't exist or errors that you handle in specific routes
 require("./error-handling")(app);
